Move loading to new file and add pyquil loader
